Update python-openstackclient spec
Because all other openstack commands use 'set' instead of 'update'
as mentioned in [1] by gongysh and also in the implemetation [1],
the 'set' is used for 'openstack vim set', this patch change the
'update' to 'set' in the spec of osc commands.

This patch also update contributors based on the history of
openstack review system.

[1] https://review.openstack.org/#/c/536032/

Change-Id: I61716c48b31cfa07ed6ccd24aa3e480ff083394a
Implements: blueprint tacker-support-python-openstackclient