Update Spec: Add support for VNF packages
The node template name of type tosca.nodes.nfv.Vdu.Compute fulfills the
purpose of the "SwImageDescId" attribute of the SwImageDesc information
element in ETSI GS NFV-IFA 011 [1] and hence it will be used in APIs to
identify the software image id from the VNFD perspective. Please refer
to clauses 6.8.3.7 and 6.8.4.7 of ETSI GS NFV-SOL 001 V2.6.1 [1].
During processing of vnf package, we can read the software images and
set software_image_id with the name of the node template name.

For example:
    VDU1:
      type: tosca.nodes.nfv.Vdu.Compute
      properties:
        ..
        sw_image_data:
          name: Software of VDU1
          version: '0.4.0'
          checksum:
            algorithm: sha-256
            hash: b9c3036539fd7a5f87a1bf38eb05fdde8b556a
          container_format: bare
          disk_format: qcow2
          min_disk: 1 GB
          size: 1 GB
      artifacts:
        sw_image:
          type: tosca.artifacts.nfv.SwImage
          file: Files/images/cirros-0.4.0-x86_64-disk.img

In this case, software_image_id will be set to VDU1.

A VNF can contain multiple deployment flavours and when VNF will be
instantiated, flavour will be passed in the request.
Added `vnf_deployment_flavors` db table so that we can validate the request
directly from the information stored in the db table instead of parsing CSAR
file again and again. Also, changed the column names from 'flavor*' to
'flavour*'.

[1] : https://www.etsi.org/deliver/etsi_gs/NFV-SOL/001_099/001/02.06.01_60/gs_NFV-SOL001v020601p.pdf

Change-Id: I1a6e7d69103b75a156c67cc407afb54ce04dff6d