Support Heal operation for CNF
This spec proposes the Heal operation for CNF based on ETSI NFV-SOL
002/003 v2.6.1. For ETSI NFV-SOL002, to re-create a VNFC mapped to Pod
is supported. For ESTI NFV-SOL003, to respawn the VNF instance is
defined as the Heal operation. The following changes are described:

For ETSI NFV-SOL002:
* Kubernetes InfraDriver stores the VnfcResourceInfo data type in
the Instantiate operation for CNF.
* Kubernetes InfraDriver validates the resource type and calls
Kubernetes Python-client to delete a Pod. The new Pod is re-created
by Kubernetes.

For ETSI NFV-SOL003:
* No change is required.

Change-Id: I9f9abcc277945476488d34fe165bc22a60c15e76
Blueprint: support-cnf-heal