Support Healing Kubernetes Master/Worker-nodes with Mgmtdriver
In this spec, we propose the Heal operation for Kubernetes Master
or Worker-nodes when the VNF is composed of Kubernetes cluster.

Proposed changes for the Heal in ETSI-NFV-SOL002:
* Mgmtdriver supports ``heal_start`` for healing worker nodes to
evacuate Pods and ``heal_end`` for worker and master nodes to join
the existing Kubernetes cluster.
* A sample script is provided to configure Kubernetes cluster.

Proposed changes for the Heal in ETSI-NFV-SOL003:
* None.

Change-Id: Ia9d49a724f4874a7ac632db97d1e417d91af37ce
Blueprint: mgmt-driver-for-k8s-heal