Support Multi-tenant for Kubernetes VIM
This spec enables Users to deploy their CNF on Kubernetes VIM
toward target namespace specified in its request.
Appropriate user accounts with authorization to the namespace
should be created in Kubernetes before the operations.

Tacker specifies the target namespace from MCIOP as Kubernetes
resource files provided in VNF Package first, and update it
with the one in the request. When no target namespace is provided,
"default" is chosen.

Co-authored-By: Masaki Ueno <masaki.ueno.up@hco.ntt.co.jp>
Implements: blueprint k8s-namespace
Change-Id: I92cdd45f8bd0ae0db8952e45ad827ffb96fa42f2