Support PVs for k8s cluster with MgmtDriver
The Tacker Wallaby release supports to manage a Kubernetes cluster as
VNF instance with ETSI NFV-SOL003 VNF Lifecycle Management
interface with MgmtDriver according to the spec "Support deploying
Kubernetes cluster with MgmtDriver".
This specification proposes that the deployed Kubernetes cluster VNF
has a Storage server with Cinder volume to enable Users to deploy CNF
which has PersistentVolume on it.

Implements: blueprint pv-k8s-cluster
Change-Id: I4be534d4fb94b3ba2feb4718603bcb8dd42ef69e