Support VNF update operations using MgmtDriver
Currently there is no preamble or postamble supported in the MgmtDriver
for the Modify VNF operation, so we will support configuration changes
to VNF instances by adding preamble and postamble to MgmtDriver.
There are many use cases for Modify VNF operations, but this
specification focuses on changing the configuration of ConfigMap and
Secret in Kubernetes and changing the image parameters in the Pod and
Deployment manifests.

Implements: blueprint container-update
Change-Id: I540212b357241db8d2f86c8d0a7305a56464fffb