Enhance Placement Process
This specification supports the placement functionality for VNF instance
and enhances the applicability of Tacker to various systems.
The placement constraints are defined in ETSI NFV-SOL 003 v3.3.1
and that VNFM sends to NFVO in order to the resource placement decision.
In VNF Lifecycle Management (LCM), there are some cases that VNFs are
not deployed due to placement constraints or
lack of availability zone's resources.
From a high-availability perspective, Tacker needs to implement a
fallbackBestEffort option in Grant Request and availability zone
reselection functions in Placement Process.

Implements: blueprint enhance-placement
Change-Id: I654cc26f30dda1cc756ddee790a16f5855ea4443