Add 2023.1 specs directory
Using release number here. In Zed cycle, TC passed a resolution[1] and
updated the release Identification document[2] to use the release
number as primary identifier in the development cycle. Release name
will be used in marketting and release team tooling (until they are
migrated to work with release number) only.

[1] https://governance.openstack.org/tc/resolutions/20220524-release-identification-process.html
[2] https://governance.openstack.org/tc/reference/release-naming.html

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: I9c56ef2a1bbce716a37cac83e4316e37d989b386