SRBAC:Implement project personas in Tacker
In the OpenStack Zed cycle, the Technical committee proposed
introducing the project scope reader and member role [1].

A project-reader persona will operate within its project
resource, and have read-only access within the project.

A project-member is denoted by someone with a member role
on a project. It is intended to be used by end users who
consume resources within a project. It inherits all the
permissions of a project-reader.

In order to implement the project-member and project-reader role
we need to depreciate the existing "owner" rule which allows
any role to behave as the owner of the project.

This specification discusses the implementation of the project
personas (project-reader and project-member) in Tacker.

[1] https://review.opendev.org/c/openstack/governance/+/847418

Implements: bp implement-project-personas
Change-Id: I12b8168f14d1e0286df34a06c4f0020365e66308